+++ Backend/Services/ItemService.cs
@@ -8,10 +8,12 @@ public class ItemService : IItemService
 {
     private readonly IItemRepository _repository;
     private readonly ILogger _logger;
+    private readonly IUserBusinessIdResolver _claimsResolver;
     
-    public ItemService(IItemRepository repo, ILogger<ItemService> logger)
+    public ItemService(IItemRepository repo, IUserBusinessIdResolver claimsResolver, ILogger<ItemService> logger)
     {
         _repository = repo;
+        _claimsResolver = claimsResolver;
         _logger = logger;
     }
 
@@ -19,7 +21,7 @@ public class ItemService : IItemService
     {
         try
         {
-            item.BusinessId = GetBusinessIdFromClaims(user);
+            var businessId = _claimsResolver.GetBusinessIdFromClaimsPrincipal(user);
             _repository.Add(item);
         }
         catch (Exception e)
@@ -38,7 +40,7 @@ public class ItemService : IItemService
             // This call throws an exception if it can't find the item.
             _repository.GetById(item.Id);
             
-            var businessId = GetBusinessIdFromClaims(user);
+            var businessId = _claimsResolver.GetBusinessIdFromClaimsPrincipal(user);
 
             if (item.BusinessId != businessId)
                 throw new Exception("You do not have permission to modify items for this business.");
@@ -64,7 +66,7 @@ public class ItemService : IItemService
     {
         try
         {
-            var businessId = GetBusinessIdFromClaims(user);
+            var businessId = _claimsResolver.GetBusinessIdFromClaimsPrincipal(user);
             
             if (businessId != item.BusinessId)
                 throw new Exception("You do not have permission to delete items for this business.");
@@ -87,7 +89,7 @@ public class ItemService : IItemService
             // This call throws an exception if it can't find the item.
             var item = _repository.GetById(id);
             
-            var businessId = GetBusinessIdFromClaims(user);
+            var businessId = _claimsResolver.GetBusinessIdFromClaimsPrincipal(user);
             if (businessId != item.BusinessId)
                 throw new Exception("You do not have permission to delete items for this business.");
             
@@ -107,7 +109,7 @@ public class ItemService : IItemService
     {
         try
         {
-            var businessId = GetBusinessIdFromClaims(user);
+            var businessId = _claimsResolver.GetBusinessIdFromClaimsPrincipal(user);
             var globalInventory = _repository.GetAll();
             var businessItems = globalInventory.Where((item) => item.BusinessId == businessId);
 
@@ -119,30 +121,5 @@ public class ItemService : IItemService
         }
 
         return Enumerable.Empty<Item>();
-    }
-
-    // Probably should just keep the approach we had prior and use roles/scopes to check if a user can access the business they want.
-    // But for ease of implementation now, I'm opting for this approach.
-    // We can always change in the future.  Since we're not actually going to release this, I don't see any reason to make this more complicated than we actually need it to be.
-    // For ease assuring that this method works as expected, I'm marking it public so I can unit test it.
-    public uint GetBusinessIdFromClaims(ClaimsPrincipal user)
-    {
-        // Get NameIdentifer.
-        var nameIdent = user.Claims.FirstOrDefault(claim => claim.Type == ClaimTypes.NameIdentifier)?.Value;
-
-        if (nameIdent is null)
-            throw new Exception("Could not find NameIdentifier within claims.");
-
-        // Auth0 NameIdentifier should be formatted as auth0|number
-        var identNumString = nameIdent.Split("|")[1];
-        var identNumSubString = identNumString.Substring(identNumString.Length - 8);
-        
-        _logger.LogInformation($"Identity sub lenth is {identNumSubString.Length}");
-
-        var identNum = Convert.ToUInt32(identNumSubString, 16);
-
-        _logger.LogInformation($"Converted Id is: {identNum}");
-
-        return identNum;
     }
 }
